Third Commit

+++ my_blog/article/views.py
@@ -6,36 +6,78 @@ from django.http import HttpResponse
 from .forms import ArticlePostForm
 # 引入User模型
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+# 引入分页模块
+from django.core.paginator import Paginator
 
 from .models import ArticlePost
 import markdown
 
+# 引入 Q 对象
+from django.db.models import Q
+
 # Create your views here.
 # 文章列表
 def article_list(request):
-    # 取出所有博客文章
-    articles = ArticlePost.objects.all()
+    search = request.GET.get('search')
+    order = request.GET.get('order')
+    # 用户搜索逻辑
+    if search:
+        if order == 'total_views':
+            # 用 Q 对象进行联合搜索
+            article_list = ArticlePost.objects.filter(
+                Q(title__icontains=search) | 
+                Q(body__icontains=search)
+            ).order_by('-total_views')
+        else:
+            article_list = ArticlePost.objects.filter(
+                Q(title__icontains=search) | 
+                Q(body__icontains=search)
+            )
+    else:
+        # 将 search 参数重置为空
+        search = ''
+        if order == 'total_views':
+            article_list = ArticlePost.objects.all().order_by('-total_views')
+        else:
+            article_list = ArticlePost.objects.all()
+    
+    # 每页显示 1 篇文章
+    paginator = Paginator(article_list, 2)
+    # 获取 url 中的页码
+    page = request.GET.get('page')
+    # 将导航对象相应的页码内容返回给 articles
+    articles = paginator.get_page(page)
+
     # 需要传递给模板的对象
-    context = { 'articles': articles }
+    context = { 'articles': articles, 'order': order, 'search': search }
     return render(request, 'article/list.html', context)
 
 # 文章列表
 def article_detail(request, id):
     # 取出相应的文章
     article = ArticlePost.objects.get(id = id)
+
+    # 浏览量 + 1
+    article.total_views += 1
+    article.save(update_fields=['total_views'])
     # 需要传递给模板的对象
     # 将markdown 语法渲染成html样式
-    article.body = markdown.markdown(article.body, extensions = [
+    md = markdown.Markdown(extensions = [
         # 包含 缩写、表格等常用拓展
         'markdown.extensions.extra',
         # 语法高亮拓展
         'markdown.extensions.codehilite',
+        # 目录拓展
+        'markdown.extensions.toc'
     ])
-    context = { 'article': article }
+    article.body = md.convert(article.body)
+    context = { 'article': article, 'toc': md.toc}
     # 载入模板，并返回context对象
     return render(request, 'article/detail.html', context)
 
 # 写文章
+@login_required(login_url='/userprofile/login/')
 def article_create(request):
     # 判断用户是否提交数据
     if request.method == 'POST':
@@ -45,8 +87,8 @@ def article_create(request):
         if article_post_form.is_valid():
             # 保存数据，但暂时不提交到数据库中
             new_article = article_post_form.save(commit=False)
-            # 指定数据库中 id=1 的用户为作者
-            new_article.author = User.objects.get(id=1)
+            # 指定目前登录的用户为作者
+            new_article.author = User.objects.get(id=request.user.id)
             # 将新文章保存在数据库中
             new_article.save()
             # 完成后返回文章列表
@@ -64,10 +106,13 @@ def article_create(request):
         return render(request, 'article/create.html', context)
 
 # 删除文章
+@login_required(login_url='/userprofile/login/')
 def article_safe_delete(request, id):
     if request.method == 'POST':
         # 根据 id 获取需要删除的文章
         article = ArticlePost.objects.get(id=id)
+        if request.user != article.author:
+            return HttpResponse('抱歉，你无权修改这篇文章！')
         # 调用 .delete() 方法删除文章
         article.delete()
         # 完成删除后返回文章列表
@@ -76,9 +121,13 @@ def article_safe_delete(request, id):
         return HttpResponse('仅允许post请求')
 
 # 更新文章
+@login_required(login_url='/userprofile/login/')
 def article_update(request, id):
     # 获取需要修改的具体文章对象
     article = ArticlePost.objects.get(id=id)
+
+    if request.user != article.author:
+        return HttpResponse('抱歉，你无权修改这篇文章！')
     # 判断用户是否为 POST 提交表单数据
     if request.method == 'POST':
         # 将提交的数据赋值到表单实例中
